enhanced_steam_bot/requirements.txt
aiohttp>=3.9.0
aiofiles>=23.2.0
Pillow>=10.0.0
python-dotenv>=1.0.0
apscheduler>=3.10.0
google-generativeai>=0.8.0
openai>=1.50.0
anthropic>=0.39.0
asyncpg>=0.29.0
tenacity>=8.2.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
structlog>=23.2.0
rich>=13.7.0
httpx>=0.27.0
